moved/renamed test files
