+++ src/test/java/UnitTest/Utils/PasswordHashTest.java
@@ -1,5 +1,6 @@
-package Utils;
+package UnitTest.Utils;
 
+import Utils.PasswordHash;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
